# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    ActionOrigin:
      type: string
      enum:
      - CODE
      - CONFIG
...
    Response37:
      type: object
      properties:
        exitCode:
          type: integer
          format: int32
        stdout:
          type: string
        stderr:
          type: string
    Response38:
      type: object
      properties:
        actions:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceActionDto"
...
    WorkspaceActionDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        origin:
          $ref: "#/components/schemas/ActionOrigin"
        interactive:
          type: boolean
        runnable:
          type: boolean
...
paths:
...
  /api/repositories/{repoId}/workspaces/{workspaceId}/actions:
    get:
      parameters:
      - name: repoId
        in: path
        required: true
        schema:
          type: string
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response38"
      summary: List
      tags:
      - Workspace Actions Controller
  /api/repositories/{repoId}/workspaces/{workspaceId}/actions/{actionId}/run:
    post:
      parameters:
      - name: actionId
        in: path
        required: true
        schema:
          type: string
      - name: repoId
        in: path
        required: true
        schema:
          type: string
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response37"
      summary: Run
      tags:
      - Workspace Actions Controller
...
info:
  title: qits API
  version: 1.0.0-SNAPSHOT