# service/src/main/webui/src/app/api/model/actionOrigin.ts
/**
 * qits API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ActionOrigin = {
    Code: 'CODE',
    Config: 'CONFIG'
} as const;
export type ActionOrigin = typeof ActionOrigin[keyof typeof ActionOrigin];